Buffer the current work on vfs. Add some commands.

+++ src/dtb.c
@@ -21,8 +21,8 @@ void fdt_traverse(void (*callback)(fdt_prop *, char *, char *))
     if (bswap_32(dtb_address->magic) != FDT_MAGIC)
         return;
 
-    uint32_t *struct_sp = (uint32_t *) ((char *)dtb_address + bswap_32(dtb_address->off_dt_struct));
-    char *string_sp = (char *) ((uint32_t) dtb_address + bswap_32(dtb_address->off_dt_strings));
+    uint32_t *struct_sp = (uint32_t *) ((char *) dtb_address + bswap_32(dtb_address->off_dt_struct));
+    char *string_sp = ((char *) dtb_address + bswap_32(dtb_address->off_dt_strings));
     
     char *node_name = NULL;
     bool END = false;
